Move startServer and closeServer to own file and remove server/server folder
